drawing tool flow
