+++ src/containers/navigation/location-tools/index.tsx
@@ -1,32 +1,24 @@
-import { useCallback, useEffect, useMemo } from 'react';
-
-import { useMap } from 'react-map-gl';
-
-import { useRouter } from 'next/router';
+import { useCallback, useEffect } from 'react';
 
 import cn from 'lib/classnames';
 
 import { analysisAlertAtom, analysisAtom, skipAnalysisAlertAtom } from 'store/analysis';
-import { drawingToolAtom, drawingUploadToolAtom } from 'store/drawing-tool';
 import { activeGuideAtom } from 'store/guide';
 import { locationsModalAtom } from 'store/locations';
-import { mapCursorAtom } from 'store/map';
-import { mapSettingsAtom } from 'store/map-settings';
-import { printModeState } from 'store/print-mode';
 import { locationToolAtom } from 'store/sidebar';
 
-import { useRecoilState, useSetRecoilState, useResetRecoilState, useRecoilValue } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 
 import AnalysisAlert from 'containers/alert';
+import WidgetDrawingTool from 'containers/datasets/drawing-tool';
+import WidgetDrawingUploadTool from 'containers/datasets/drawing-upload-tool';
 import Helper from 'containers/guide/helper';
 import LocationDialogContent from 'containers/location-dialog-content';
 
 import { Dialog, DialogTrigger } from 'components/dialog';
 import Icon from 'components/icon';
 
-import AREA_SVG from 'svgs/sidebar/area.svg?sprite';
 import GLASS_SVG from 'svgs/sidebar/glass.svg?sprite';
-import UPLOAD_SVG from 'svgs/sidebar/upload.svg?sprite';
 
 const MANGROVES_SKIP_ANALYSIS_ALERT = 'MANGROVES_SKIP_ANALYSIS_ALERT';
 
@@ -38,21 +30,6 @@ const LocationTools = () => {
   const [skipAnalysisAlert, setSkipAnalysisAlert] = useRecoilState(skipAnalysisAlertAtom);
   const guideIsActive = useRecoilValue(activeGuideAtom);
 
-  const setDrawingToolState = useSetRecoilState(drawingToolAtom);
-  const setDrawingUploadToolState = useSetRecoilState(drawingUploadToolAtom);
-  const resetAnalysisState = useResetRecoilState(analysisAtom);
-
-  const resetMapCursor = useResetRecoilState(mapCursorAtom);
-
-  const isPrintingMode = useRecoilValue(printModeState);
-  const isPrintingId = isPrintingMode ? 'print-mode' : 'no-print';
-
-  const { [`default-desktop-${isPrintingId}`]: map } = useMap();
-  const resetMapSettingsState = useResetRecoilState(mapSettingsAtom);
-  const { asPath, replace } = useRouter();
-
-  const queryParams = useMemo(() => asPath.split('?')[1], [asPath]);
-
   const openLocationsModal = useCallback(() => {
     if (!locationsModalIsOpen) setLocationsModalIsOpen(true);
   }, [locationsModalIsOpen, setLocationsModalIsOpen]);
@@ -63,66 +40,6 @@ const LocationTools = () => {
       saveLocationTool(null);
     }
   }, [isAnalysisAlertOpen, setLocationsModalIsOpen, saveLocationTool]);
-
-  const handleDrawingToolView = useCallback(() => {
-    setDrawingToolState((drawingToolState) => ({
-      ...drawingToolState,
-      showWidget: true,
-      enabled: false,
-    }));
-    setDrawingUploadToolState((drawingUploadToolState) => ({
-      ...drawingUploadToolState,
-      showWidget: false,
-      enabled: false,
-    }));
-
-    resetAnalysisState();
-    resetMapSettingsState();
-    resetMapCursor();
-    saveLocationTool('area');
-
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    replace(`/custom-area${queryParams ? `?${queryParams}` : ''}`, null);
-  }, [
-    setDrawingToolState,
-    resetAnalysisState,
-    resetMapCursor,
-    replace,
-    queryParams,
-    saveLocationTool,
-    resetMapSettingsState,
-    setDrawingUploadToolState,
-  ]);
-
-  const handleDrawingUploadToolView = useCallback(() => {
-    setDrawingUploadToolState((drawingUploadToolState) => ({
-      ...drawingUploadToolState,
-      showWidget: true,
-      enabled: false,
-    }));
-    setDrawingToolState((drawingToolState) => ({
-      ...drawingToolState,
-      showWidget: false,
-      enabled: false,
-    }));
-
-    resetAnalysisState();
-    resetMapSettingsState();
-    resetMapCursor();
-    saveLocationTool('upload');
-
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    replace(`/custom-area${queryParams ? `?${queryParams}` : ''}`, null);
-  }, [
-    resetAnalysisState,
-    resetMapCursor,
-    replace,
-    queryParams,
-    saveLocationTool,
-    resetMapSettingsState,
-    setDrawingUploadToolState,
-    setDrawingToolState,
-  ]);
 
   const openAnalysisAlertModal = useCallback(() => {
     setAnalysisAlert(true);
@@ -187,61 +104,21 @@ const LocationTools = () => {
 
           <LocationDialogContent close={closeMenu} />
         </Dialog>
-        {/* //*DRAW AREA* */}
-        <button
-          type="button"
-          className="flex w-28 cursor-pointer items-center justify-center"
-          onClick={handleDrawingToolView}
-          data-testid="drawing-tool-button"
-        >
-          <Helper
-            className={{
-              button: '-bottom-3.5 right-2.5 z-[20]',
-              tooltip: 'w-fit-content max-w-[400px]',
-            }}
-            tooltipPosition={{ top: -65, left: 0 }}
-            message="use this function to calculate statistics for your own custom area of interest"
-          >
-            <div className="flex flex-col items-center space-y-1">
-              <Icon
-                icon={AREA_SVG}
-                className={cn({
-                  'h-8 w-8 rounded-full fill-current text-white': true,
-                })}
-                description="Area"
-              />
-              <p className="whitespace-nowrap font-sans text-sm text-white">Draw area</p>
-            </div>
-          </Helper>
-        </button>
 
-        {/* //*UPLOAD SHAPEFILE* */}
+        {/* //*DRAW AREA* */}
         <Helper
           className={{
-            button: '-bottom-3.5 right-9 z-[20]',
+            button: '-bottom-3.5 right-2.5 z-[20]',
             tooltip: 'w-fit-content max-w-[400px]',
           }}
           tooltipPosition={{ top: -65, left: 0 }}
-          message="use this button to upload an area you want to analyze."
+          message="use this function to calculate statistics for your own custom area of interest"
         >
-          <button
-            type="button"
-            className="flex w-28 cursor-pointer items-center justify-center"
-            onClick={handleDrawingUploadToolView}
-            data-testid="worldwide-button"
-          >
-            <div className="flex flex-col items-center space-y-1">
-              <Icon
-                icon={UPLOAD_SVG}
-                className={cn({
-                  'h-8 w-8 fill-current text-white': true,
-                })}
-                description="Upload"
-              />
-              <p className="whitespace-nowrap font-sans text-sm text-white">Upload shapefile</p>
-            </div>
-          </button>
+          <WidgetDrawingTool />
         </Helper>
+
+        {/* //*UPLOAD SHAPEFILE* */}
+        <WidgetDrawingUploadTool />
       </div>
       <AnalysisAlert />
     </div>
