+++ src/store/upload-file/index.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const uploadFileAtom = atom<boolean>({
+  key: 'upload-file-error',
+  default: false,
+});
